GUVNOR-804 + GUVNOR-686: grid based on celltable prototype
git-svn-id: https://svn.jboss.org/repos/labs/labs/jbossrules/trunk@35958 c60d74c8-e8f6-0310-9e8f-d4a2fc68ab70

+++ drools-repository/src/main/java/org/drools/repository/AssetItemIterator.java
@@ -60,7 +60,7 @@ public class AssetItemIterator
     /**
      * @param i The number of rules to skip.
      */
-    public void skip(int i) {
+    public void skip(long i) {
         it.skip( i );
     }
